#28 - landing page (#39)
* #28 Initial setup for landing page

Include installing tailwind and dependencies

* #28 Tailwind fix

As in title. There was an error with tailwind version

* #28 Test work for now

Commiting so i can check everything on my pc

* #28 Landing page demo version

As in title. For now without any special funcionality

* #28 Landing page almost finished

Page almost finished from design site, of course without any special funcionality, it will be added in the future.

* #28 Fixed Landing page placing

As in title. Also minor typos

* #28 Alias fix

As in title.

* #28 Added polygon on second image

As in title.

Co-authored-by: Łukasz Podlipski <73656796+LukaszPodlipski@users.noreply.github.com>