#26 - main page (#73)
* #28 Initial setup for landing page

Include installing tailwind and dependencies

* #28 Tailwind fix

As in title. There was an error with tailwind version

* #28 Test work for now

Commiting so i can check everything on my pc

* #28 Landing page demo version

As in title. For now without any special funcionality

* #28 Landing page almost finished

Page almost finished from design site, of course without any special funcionality, it will be added in the future.

* #28 Fixed Landing page placing

As in title. Also minor typos

* #26 Created Most Components of main-page

As in title

* #26 Updated Mainpage progress

As in title

* #26 Cleaned Home and added funcionality to navbar

Cleaned up unnecessary code from tailwind UI components. Added scrolling into navbar

* #26 Added Map and demo of list

As in title

* #26 fixed package log

as in title

* #26 Fixed navbar

As in title, also chnged eslint.js options to fix problems, updated package.json with npm

* #26 Worked on datafield list

As in title

* #26 Changes and updates

- reset and save buttons made correctly
- map/list section connected with viewport height
- deleted regulations section, i think that better option is to download reg, reading on site is problematic and not comfortable
- fixed problem with under lighting nav text

* #26 update

as in title

* #26 Fixed some minor vue bug

As in title. It was some dependency that had to be reinstalled to newer version

* #26 Created filter section with most of features, fixed some bugs

As in title

* #26 Working npm build

There were some problems with build, but this config and packages work for now

* #26 Changed files structure and updated filter section

as in title

* #26 Updated main page work

Styled for desktop and mobile,
changed filter section,
fixed some bugs,
added new features

* #26 Text overflow animation fixed

As in title.

* #26 Updated

changed datafield appearance,
fixed bug with filter section, made company list easly scrollable on mobile

* #26 Cooperate font fix

As in title, also fixed scrolling on hover

* #26 Changed css to scss

Refactorized code, deleted unused classes, corrected class names

* #26 Responsive text sizes

As om title.

* #26 updated map code

as in title

* #26 Reset filters button funcionality

As in title.

* #26  Cleanup

As in title.

* #26 Cleared env tokens

As in title.

* #26 Moved mapbox token to .env

As in title

* #26 updeted styles for ultrawide displays and made map tokens private

As in title

* #26 Min-width set to container on pc view

As in title.

* #26 Updates connected with comments

As in title

* #26 Alias fix

As in title.

* #26 Changed static tags to test tags loaded from json file

As in title.

* #26 Changed jquery to conditional styles in vue

As in title.

* #26 Moved all img and svg to assets/images folder

As in title.

* #74 New design of datafield

As in title.

* #26 Fixes connected with contributors comments

As in title

* #26 Fixes after merging

As in title

* #26 Small fix to div wrapping

As in title.

* #26 Removed empty lines

As in title.

* #26 Few changes with styling

As in title

* #26 Custom scrollbar

As in title.

* #26 Clean up and Composition Api change

As in title.

* #26 Updates connected with review comments

- created function isMobile() in separated file
- made some classes and functions names more descriptive
- removed unused datafiles for mobile and desktop - now we have one datafield component. On desktop it displays tags, on mobile no.

* #26 Transition from jquery to conditional styling

As in title.

* #26 V-for usage in test data

As in title.

* #26 More descriptive name

As in title.

* #26 Added global scss variables for colors

As in title

Co-authored-by: brzozka <bartosz.brzozka@blumilk.pl>
Co-authored-by: brzozka <39692087+brzozka@users.noreply.github.com>
Co-authored-by: gLenczuk <lenczuk.dev@gmail.com>