#127 - text scrolling on hover in company names (#130)
* #101 Working fetching from faculties.json

As in title. Also deleted one config that disrupted correct funcionality of docker web

* #101 deleted old styling

As in title.

* #101 Working data with static path

Changed to composition api and got data from backend, with static path for now.

* #101 Changed static data to json in file and DataFields

For now without correct funcionality.

* #101 Working data with store usage

As in title.

* #101 Change component fetch to vuex store

As in title.

* #101 Active tags using vuex

Also removed ununsed mitt, for now reset only works on tags.

* #101 Setup for merged state

As in title.

* #101 Created companyMerged and changed all static data to data from json

As in title.

* #101 removed unused file.

As in title.

* #101 Removed unnecessary files and code

As in title.

* #101 Changes to naming/ added wrapper to fetch

As in title.

* #101 Changed import name

As in title.

* #101 Removed redundant async

As in title.

* #101 small fixes

As in title.

* #127 Text scrolling fixed

As in title.